evix: tighten config boundaries and daemon structure
