+++ crates/evix-daemon/Cargo.toml
@@ -8,15 +8,16 @@ license.workspace      = true
 repository.workspace   = true
 
 [dependencies]
-anyhow.workspace        = true
-evix.workspace          = true
-evix-protocol.workspace = true
-futures-util.workspace  = true
-libc.workspace          = true
-pound.workspace         = true
-serde_json.workspace    = true
-tokio.workspace         = true
-tracing.workspace       = true
+anyhow.workspace             = true
+evix.workspace               = true
+evix-protocol.workspace      = true
+futures-util.workspace       = true
+libc.workspace               = true
+pound.workspace              = true
+serde.workspace              = true
+serde_json.workspace         = true
+tokio.workspace              = true
+tracing.workspace            = true
 tracing-subscriber.workspace = true
 
 [[bin]]
